Rename BaseEnum to AbstractEnum
