+++ UPGRADE-2.1.md
@@ -1,0 +1,6 @@
+Upgrade from 2.0 to 2.1
+=======================
+
+### Enum class rename
+
+* `BaseEnum` is now deprecated in favor of `AbstractEnum`.
